API BREAK: Remove `const char* codec' wrapper functions.
Consumers are now expected to provde a SquashCodec instance instead of
passing around the codec name as a string.  This should provide a small
performance gain, encourage better usage patterns, and simplify the
API.  The disadvantage is that the library is now slightly more
difficult to use.

This also gets around the issue of providing a huge number of
character function variants for internationalized Windows applications.