Add wchar_t variants of most functions which deal with strings.
This is almost exclusively aimed at i18n on Windows, but it should also
make improve the situation on Linux when the locale is not UTF-8.  The
remaining functions which you might expect to have wide-character
versions mostly deal exclusively with ASCII strings (the codec and
plugin names).

To enable, use the "wide-char-api" cargo feature