Allow plugins to request a wrapper for the decompressed size.
Some libraries (BriefLZ and LZSSE, probably others) require knowledge
of the decompressed size in order to decompress successfully.  We
previously forced such libraries to add their own wrapper to add the
decompressed size to the compressed data, but with this patch they can
simply request that Squash do it for them.